gwikixml-attachfile script to attach files remotely to wiki.
Uses the function xmlrpc_attachfile as follows:

from graphingwiki.editing import xmlrpc_attachfile
[...]
result = xmlrpc_attachfile(wiki, page, fname, content, username, password)

where wiki is the url to wiki, page is the pagename,
fname is the name desired for the attached file,
content is the file content (in binary, it'll be encoded properly),
username and password are the credentials

NB: MetaTableEdit is undergoing changes and not quite functional in
this revision, and the functionality of MetaTable has been changed to
also include typed links to the table in addition to metadata.

Small fixes includes also.