Fixed an evil bug in savegraphdata:
 koo:: somestring SomeLink somestring

saved as SomeLink. Initially I thought this was a funny functionality
and let it be, but it actually means that MetaEdit and SetMeta could
not edit it. So, requiring complete matches for links currently.

REQUIRES REHASH

Implemented #88, #89, #145, fixed #150. MetaEdit enhanced greatly.

PLEASE WATCH OUT: EMPTYING A KEY IN METAEDIT DELETES THE KEY ALONG
WITH ALL ITS VALUES.

Some editing related bugs fixed. MetaTable now considers int, float
and ip address values when ordering the table.

GetMeta, SetMeta and gwiki-xml-meta behaviour changed with regard to
multivalued attributes. This is to enable insertion of multiple values
to the same key and to enhance interface ortogonality, kind of.

Before:

Page;key
Koo;1,2,3

in GetMeta

['key']
['Koo', [1,2,3]]

Now:

Page;key;key;key
Koo;1;2;3

in GetMeta

[['key', 'key', 'key']]
['Koo', [1,2,3]]

Please tell me if this produces difficulties and we can talk about
changing it.