Fixes #143, #98, #140, #153, #152.
Calling metatable without specifying pages should be considerably
faster because Moin's page index is not used anymore.

Adds two new macros.

[[MetaRevision(page)]]

shows how the metadata of the page in question has developed through
page revisions.

[[TagCloud(metatable-args, maxTags=x)]]

shows a tag cloud of the meta keys of the pages in question. The pages
to use can be selected with arguments similar to those of metatable,
the maxTags-argument constraints the number of tags in the cloud,
defaulting at 50.

These macros should be considered experimental for now. Enjoy.