fixed #128, #108, #130, #172 and #173
Other small superfixes includes also.

Meta values beginning with a link

 description:: FrontPage is a Wiki page

were erroneusly saved as links in a manner that disabled their
automated handling. This is actually a corner case of the bug fixed in
changeset [218].

The fix to this issue does not probably warrant a rehash as it is
(hopefully) a relatively rare case, and I'm planning for some bigger
changes in the relatively near future that would require a rehash
anyway. Resaving any capricious pages will probably suffice for the
time being.