Fixes #387, #388, #142
Already open shelves can be used by using the getgraphdata-function in
patterns.py. This puts a ReadLock into effect, and using any of the
process_edit toolchain functions will save the page data with
savegraphdata and re-instate the ReadLock after doing that. getgraphdata
also tries to ensure that locks are closed at the end of the request -
there are no such guarantees with GraphData.

Using this in macros seems to have adverse effects, though,
and is not encouraged - if you find out exactly why this is, please let
me know.

Please test the fix to #142.