Copying request causes severe locking problems. Added a hack to alleviate the problem. The hack should be removed asap and the copying practice stopped altogether and the offending pieces of code refactored. Otherwise we'll pay later.
The problem is caused by copying functions' intractions with the
magical .graphdata-attribute. For example InlineGraph does such
copying, which caused simultaneous read and write locks to appear when
saving a page with inlinegraphs. This because there were two graphdata
objects doing locking. It in turn brought whole wiki to a halt.

--HG--
branch : moin-1.6-branch