Tries to fix #570.
Additionally, restricts metadata values to single line. MoinMoin
considers that the in the following markups, the lines following the
definitions are indented at the same level as the definition list:

 a:: 1

 * list

 a:: 2

 1. list

 a:: 3

 {{{preformatted}}}

 a:: 4

 just some text

Previously, Gwiki 1.6-branch has followed this thinking, though not in a
fully consistent manner (empty lines and line breaks were omitted,
potentially wrecking havoc).

This was not the expected behaviour to many users. Due to this, this
revision changes the behaviour so that only the text that is on the same
line as the meta key is accepted as the meta value.

This may be subject to change in the future based on usage scenarios
and user input.

REQUIRES REHASH

--HG--
branch : moin-1.6-branch