Implements #169, #110
Meta indirection support, and inlink data to meta.

For example, if you have pages A, B and C such as

A

{{{
 bar:: [[B]]
}}}

B

{{{
 foo:: [[C]]
}}}

C

{{{
[[A]]
 baz:: grue
}}}
and try MetaTable

{{{
<<MetaTable(A, ||gwikiinlinks||bar->foo||bar->foo->baz||bar->gwikiinlinks||)>>
}}}
you should get the table

{{{
A | [[C]] | [[C]] | grue | [[A]] |
}}}

Both are implemented in get_metas, and therefore should work in most of
gwiki in addition to metatable. Indirection and inlinks attrs should
also be immutable, so that you cannot edit them - they are filtered in
editing, and in MetaEdit also when the form is displayed.

(replaying exec's commit from old svn repo)