Fixes #751, #753 Implements styles for graph ordernodes. Tries to simplify the graph interface a bit (anonymous graph -> filter this page etc)
Re-introduces property-pages for meta keys.

Implements #649

Brief description on the implementation

Suppose you have create pages with MetaFormEdit as follows

{{{
<<NewPage(TestTemplate, Create Page,,,EditMetaForm)>>
}}}

and the TestTemplate is the following:

{{{
= @PAGE@ =

 aa::
 bb::
----
<<LinkedIn>>
----
CategoryTest
}}}

By default, the metaformedit action works as before - i.e. there will be a
selection form element listing the existing values in the wiki as well
as a text box where you can enter additional values.

However, you can give properties to metakeys by using their property pages
(surprise!). This will in effect tell MetaFormEdit what kinds of selectors
to use. The meta keys currently used on the property pages are hint and
constraint (description is also defined as descriptive text for the meta
key, but not used at the moment).

The currently supported values for hint are:
 * selection, checkbox, radio, textbox, textarea

The values textbox and textarea will never result in additional textboxes
in the form (the "where you can enter additional values" as described above).

The currently supported values for constraint are:
 * existing

aaProperty

{{{
 constraint::
 description:: Selections on this meta use radio buttons
The hint=radio tells MetaFormEdit to use radio buttons with existing values
 hint:: radio
}}}

bbProperty

{{{
 description:: This is a checkbox selection meta.
 hint:: checkbox
The constraint=existing means that there will be no text box in addition to the checkboxes -
you have to select from values already existing on some wiki pages. It's considered a good
practice to have preset pages for desired values.
 constraint:: existing
}}}