Fixed #759
Actually cases 594, 596, 759 seem to have been regressions on the same
issue. Hope this nails the issue of handling preformatted areas for good.

The issue with gwikilabel was due to special handling of the gwikilabel
values. This was due to my own requirement for setting empty labels.
Try it - make a page with the following line and show a graph

{{{
 gwikilabel::
}}}

However, this was implemented by not stripping the labels. This meant that
when you have a label {{{u'label here '}}}, the replacement logic searched
for the stripped value {{{u'label here'}}}, which resulted in duplicate
values. The label handling was modified so that the value is stripped but
empty values are still allowed. They cannot currently be edited with any
sensible means with any meta editing facilities. Maybe it should be modified
e.g. by using "gwikiempty" for an empty value.

Rehash is not required, but you'll need to manually fix pages with duplicate
labels.