Fix a small bug in including metatables with attachments as values.
Enable looking at graphs on underlay pages.

Fix including pages when formatter is not defined for some reason.

Remove some dead code.

Change the treatment of interwiki links slightly: Links matching the
interwiki syntax are saved the same way as before. However, they are
not treated as valid interwiki links in graphs if the interwiki name
is not known in the interwiki list.

UnknownWiki:SomePage

Before: would be evaluated as a link to an unknown wiki ->
./Interwiki/

After: is evaluated as text. If UnkownWiki is added to the interwiki
list, it will be evaluated as a link.

This behaviour change stems from the frequent misbehaviour with
finnish language and it's use of the semicolon with abbreviations,
such as "MIT:stä" or "UKK:hon".