Link changes in previous commits broke some existing behaviour when editing pages over xmlrpc. For example, if a link on was {{attachment:k.txt}} on page SomePage getmeta now returns {{attachment:SomePage/k.txt|k.txt}}.
This is ok when displaying the meta values in includes, but not when
reading values, editing them, and then saving them back, which would
generate unnecessary and unwanted edits. Looking at the code, it turns
out that getting the exact values of the meta values was the common
case, not to get pretty-printed data.

I added an option to get_metas on whether to format (pretty print) the
link values or not. The default is not to do any formatting.