Added cell coloring functionality to MetaTable, as well as property override ancd backup pages. As per their names, the override pages override any property pages of the keys, whereas the backup pages are only used if there are no property pages for the keys.
Colors in property pages work as follows:

 color REGEXP:: COLORCODE

REGEXP is a regular expression against which they key values will be
evaluated. COLORCODE is either an SVG colorcode or a hex value in the
form of #ffffff. Example

Page
 a:: 1
 b:: 2

aProperty
 color 1:: red

Property b page does not exist.

backProperty
 color \d+:: #0000aa

overProperty
 color 2:: black

The result of the following metatables:

<<MetaTable(Page, ||a||b||)>>

the cell containing the value 1 for key a is red, cell containing
value 2 for b is white.

<<MetaTable(Page, ||a||b||, propbackup:=back)>>

the cell containing the value 1 for key a is red, cell containing
value 2 for b is light blue.

<<MetaTable(Page, ||a||b||, propoverride:=over)>>

the cell containing the value 1 for key a is white (regexp did not
match), cell containing value 2 for b is black.