graphingwiki.patch.d.Page.py.patch: Optimised link closing to help on link-heavy pages and tables.
The code making page links unnecessary made link texts both when opening and closing links. A long chain of functions was called just to generate </a>. This patch is a shorthand to Page.link_to to skip most of that code. formatter.url is used in Page.py in highlights already, although the use of formatter is generally avoided. Another option would have been to modify link_to, link_to_raw and wikiutil.link_tag, but this patch was less invasive.